Added download notification mask
Added the download notification mask. The mask appears while downloading
the files and shows the download progress.

Added cordova plugins source to the git tree in order to track the
changes.

Changed the Downloader plugin in order to make it possible to specify
the destination directory and file name without passing them as
url parameters.