Store updates mechanism draft + MenuMensa storage
Implemented a draft for the local store updates mechanism.
The local storage is updated and maintained in the StoreFeed
controller. This new controller updates the stores used in
the application.

The new WebService method getUpdates() obtains an associative
array such as:
	{ "menumensa" => "timestamp" }

The timestamps are then used to establish if to update or not
the localstores.

MenuMensa storage makes it possible to avoid static url values
in the code. This way if the url changes the change can easily
be made on the server side.