Isolated platform dependant code
Isolated the platform dependant code. All the platform dependant
javascript code has been moved to the js/[platform] and
plugins/[platform] directories. The functions used in the app
are added during runtime and allow a separation between the
architecture dependant code and the common code.

Removed Childbrowser plugin because it is no necessary anymore.